+++ src/D2L.Bmx/Okta/Models/AuthenticateChallengeMfaOptions.cs
@@ -1,0 +1,13 @@
+using System.Text.Json.Serialization;
+namespace D2L.Bmx;
+public struct AuthenticateChallengeMfaOptions {
+	public string FactorId { get; set; }
+	public string PassCode { get; set; }
+	public string StateToken { get; set; }
+
+	public AuthenticateChallengeMfaOptions( string factorId, string passCode, string stateToken ) {
+		FactorId = factorId;
+		PassCode = passCode;
+		StateToken = stateToken;
+	}
+}
