+++ src/D2L.Bmx/Okta/Models/AuthenticateOptions.cs
@@ -1,0 +1,10 @@
+namespace D2L.Bmx;
+public struct AuthenticateOptions {
+	public string Username { get; set; }
+	public string Password { get; set; }
+
+	public AuthenticateOptions( string username, string password ) {
+		Username = username;
+		Password = password;
+	}
+}
