+++ src/D2L.Bmx/Okta/Models/OktaApp.cs
@@ -1,0 +1,7 @@
+namespace D2L.Bmx;
+public struct OktaApp {
+	public string Id { get; set; }
+	public string Label { get; set; }
+	public string AppName { get; set; }
+	public string LinkUrl { get; set; }
+}
