+++ src/D2L.Bmx/Okta/Models/OktaError.cs
@@ -1,0 +1,7 @@
+namespace D2L.Bmx;
+public struct OktaError {
+	public string ErrorCode { get; set; }
+	public string ErrorSumamry { get; set; }
+	public string ErrorLink { get; set; }
+	public string ErrorId { get; set; }
+}
