+++ src/D2L.Bmx/Okta/Models/OktaMfaFactor.cs
@@ -1,0 +1,7 @@
+namespace D2L.Bmx;
+public struct OktaMfaFactor {
+	public string Id { get; set; }
+	public string FactorType { get; set; }
+	public string Provider { get; set; }
+	public string VendorName { get; set; }
+}
