+++ src/D2L.Bmx/Okta/Models/OktaSession.cs
@@ -1,0 +1,7 @@
+namespace D2L.Bmx;
+public struct OktaSession {
+	public string Id { get; set; }
+	public string UserId { get; set; }
+	public DateTimeOffset CreatedAt { get; set; }
+	public DateTimeOffset ExpiresAt { get; set; }
+}
