+++ src/D2L.Bmx/Okta/Models/SessionOptions.cs
@@ -1,0 +1,8 @@
+namespace D2L.Bmx;
+public struct SessionOptions {
+	public string SessionToken { get; set; }
+
+	public SessionOptions( string sessionToken ) {
+		SessionToken = sessionToken;
+	}
+}
